Fix app routing
