+++ src/components/app/app.module.css
@@ -11,11 +11,5 @@
     display: flex;
     flex-direction: row;
     justify-content: center;
-}
-
-.panel {
-    width: 600px;
-    height: 100%;
-    margin-right: 8px;
-    margin-left: 8px;
+    align-items: center;
 }
