+++ src/pages/forgot-password.module.css
@@ -1,19 +1,3 @@
-.app {
-    display: flex;
-    flex-direction: column;
-    width: 100vw;
-    height: 100vh;
-}
-
-.main {
-    flex-grow: 1;
-
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-}
-
 .name {
     composes: mt-6 from global;
 }
